gomi error

+++ src/utils/free_utils.c
@@ -3,19 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   free_utils.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: vmatsuda <vmatsuda@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*   By: vmatsuda <vmatsuda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/08 22:34:44 by vmatsuda          #+#    #+#             */
-/*   Updated: 2025/07/08 22:43:55 by vmatsuda         ###   ########.fr       */
+/*   Updated: 2025/07/11 16:43:16 by vmatsuda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "so_long.h"
-#include "free_utils.h"
 
 void	free_map(t_map *map)
 {
-	int	i;
+	size_t	i;
 
 	if (!map)
 		return ;
